worker/requirements.txt
amqp==5.2.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
google-cloud-storage==2.16.0
google-cloud-pubsub==2.21.1
httpx==0.27.0
kombu==5.3.6
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
